Move query API sources to `api` directory.
