+++ tiledb/sm/c_api/tiledb.cc
@@ -188,524 +188,6 @@ capi_return_t tiledb_as_built_dump(tiledb_string_t** out) {
 /*              QUERY             */
 /* ****************************** */
 
-int32_t tiledb_query_alloc(
-    tiledb_ctx_t* ctx,
-    tiledb_array_t* array,
-    tiledb_query_type_t query_type,
-    tiledb_query_t** query) {
-  ensure_array_is_valid(array);
-
-  // Error if array is not open
-  if (!array->is_open()) {
-    auto st = Status_Error("Cannot create query; Input array is not open");
-    *query = nullptr;
-    LOG_STATUS_NO_RETURN_VALUE(st);
-    save_error(ctx, st);
-    return TILEDB_ERR;
-  }
-
-  // Error if the query type and array query type do not match
-  tiledb::sm::QueryType array_query_type;
-  try {
-    array_query_type = array->get_query_type();
-  } catch (StatusException& e) {
-    return TILEDB_ERR;
-  }
-
-  if (query_type != static_cast<tiledb_query_type_t>(array_query_type)) {
-    std::stringstream errmsg;
-    errmsg << "Cannot create query; "
-           << "Array query type does not match declared query type: "
-           << "(" << query_type_str(array_query_type) << " != "
-           << tiledb::sm::query_type_str(
-                  static_cast<tiledb::sm::QueryType>(query_type))
-           << ")";
-    *query = nullptr;
-    auto st = Status_Error(errmsg.str());
-    LOG_STATUS_NO_RETURN_VALUE(st);
-    save_error(ctx, st);
-    return TILEDB_ERR;
-  }
-
-  // Create query struct
-  *query = new (std::nothrow) tiledb_query_t;
-  if (*query == nullptr) {
-    auto st = Status_Error(
-        "Failed to allocate TileDB query object; Memory allocation failed");
-    LOG_STATUS_NO_RETURN_VALUE(st);
-    save_error(ctx, st);
-    return TILEDB_OOM;
-  }
-
-  // Create query
-  (*query)->query_ = new (std::nothrow) tiledb::sm::Query(
-      ctx->resources(),
-      ctx->cancellation_source(),
-      ctx->storage_manager(),
-      array->array());
-  if ((*query)->query_ == nullptr) {
-    auto st = Status_Error(
-        "Failed to allocate TileDB query object; Memory allocation failed");
-    delete *query;
-    *query = nullptr;
-    LOG_STATUS_NO_RETURN_VALUE(st);
-    save_error(ctx, st);
-    return TILEDB_OOM;
-  }
-
-  // Success
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_stats(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, char** stats_json) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  if (stats_json == nullptr)
-    return TILEDB_ERR;
-
-  const std::string str = query->query_->stats()->dump(2, 0);
-
-  *stats_json = static_cast<char*>(std::malloc(str.size() + 1));
-  if (*stats_json == nullptr)
-    return TILEDB_ERR;
-
-  std::memcpy(*stats_json, str.data(), str.size());
-  (*stats_json)[str.size()] = '\0';
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_config(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_config_t* config) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-  api::ensure_config_is_valid(config);
-  query->query_->set_config(config->config());
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_config(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_config_t** config) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-  api::ensure_output_pointer_is_valid(config);
-  *config = make_handle<tiledb_config_handle_t>(query->query_->config());
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_subarray_t(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const tiledb_subarray_t* subarray) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  ensure_subarray_is_valid(subarray);
-  query->query_->set_subarray(*subarray->subarray());
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_data_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    void* buffer,
-    uint64_t* buffer_size) {  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Set attribute buffer
-  query->query_->set_data_buffer(name, buffer, buffer_size);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_offsets_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint64_t* buffer_offsets,
-    uint64_t* buffer_offsets_size) {  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Set attribute buffer
-  query->query_->set_offsets_buffer(name, buffer_offsets, buffer_offsets_size);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_validity_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint8_t* buffer_validity,
-    uint64_t* buffer_validity_size) {  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Set attribute buffer
-  query->query_->set_validity_buffer(
-      name, buffer_validity, buffer_validity_size);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_data_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    void** buffer,
-    uint64_t** buffer_size) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Get attribute buffer
-  std::tie(*buffer, *buffer_size) = query->query_->get_data_buffer(name);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_offsets_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint64_t** buffer,
-    uint64_t** buffer_size) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Get attribute buffer
-  std::tie(*buffer, *buffer_size) = query->query_->get_offsets_buffer(name);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_validity_buffer(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint8_t** buffer,
-    uint64_t** buffer_size) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Get attribute buffer
-  std::tie(*buffer, *buffer_size) = query->query_->get_validity_buffer(name);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_layout(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_layout_t layout) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Set layout
-  query->query_->set_layout(static_cast<tiledb::sm::Layout>(layout));
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_set_condition(
-    tiledb_ctx_t* const ctx,
-    tiledb_query_t* const query,
-    const tiledb_query_condition_t* const cond) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR ||
-      sanity_check(ctx, cond) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Set layout
-  query->query_->set_condition(*cond->query_condition_);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_finalize(tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  // Trivial case
-  if (query == nullptr)
-    return TILEDB_OK;
-
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Flush query
-  query->query_->finalize();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_submit_and_finalize(
-    tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  // Trivial case
-  if (query == nullptr)
-    return TILEDB_OK;
-
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  query->query_->submit_and_finalize();
-
-  return TILEDB_OK;
-}
-
-void tiledb_query_free(tiledb_query_t** query) {
-  if (query != nullptr && *query != nullptr) {
-    delete (*query)->query_;
-    delete *query;
-    *query = nullptr;
-  }
-}
-
-int32_t tiledb_query_submit(tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  // Sanity checks
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  query->query_->submit();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_has_results(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, int32_t* has_results) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *has_results = query->query_->has_results();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_status(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_query_status_t* status) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *status = (tiledb_query_status_t)query->query_->status();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_type(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_query_type_t* query_type) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *query_type = static_cast<tiledb_query_type_t>(query->query_->type());
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_layout(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_layout_t* query_layout) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *query_layout = static_cast<tiledb_layout_t>(query->query_->layout());
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_array(
-    tiledb_ctx_t* ctx, tiledb_query_t* query, tiledb_array_t** array) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  // Allocate an array object, taken from the query's array.
-  *array = make_handle<tiledb_array_t>(query->query_->array_shared());
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_est_result_size(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size) {
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  auto field_name{to_string_view<"field name">(name)};
-  if (size == nullptr) {
-    throw CAPIStatusException("Pointer to size may not be NULL");
-  }
-  auto est_size{query->query_->get_est_result_size_fixed_nonnull(field_name)};
-  *size = est_size.fixed_;
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_est_result_size_var(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_off,
-    uint64_t* size_val) {
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  auto field_name{to_string_view<"field name">(name)};
-  if (size_off == nullptr) {
-    throw CAPIStatusException("Pointer to offset size may not be NULL");
-  }
-  if (size_val == nullptr) {
-    throw CAPIStatusException("Pointer to value size may not be NULL");
-  }
-
-  auto est_size{
-      query->query_->get_est_result_size_variable_nonnull(field_name)};
-  *size_off = est_size.fixed_;
-  *size_val = est_size.variable_;
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_est_result_size_nullable(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_val,
-    uint64_t* size_validity) {
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  auto field_name{to_string_view<"field name">(name)};
-  if (size_val == nullptr) {
-    throw CAPIStatusException("Pointer to value size may not be NULL");
-  }
-  if (size_validity == nullptr) {
-    throw CAPIStatusException("Pointer to validity size may not be NULL");
-  }
-
-  auto est_size{query->query_->get_est_result_size_fixed_nullable(field_name)};
-  *size_val = est_size.fixed_;
-  *size_validity = est_size.validity_;
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_est_result_size_var_nullable(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_off,
-    uint64_t* size_val,
-    uint64_t* size_validity) {
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  auto field_name{to_string_view<"field name">(name)};
-  if (size_off == nullptr) {
-    throw CAPIStatusException("Pointer to offset size may not be NULL");
-  }
-  if (size_val == nullptr) {
-    throw CAPIStatusException("Pointer to value size may not be NULL");
-  }
-  if (size_validity == nullptr) {
-    throw CAPIStatusException("Pointer to validity size may not be NULL");
-  }
-  auto est_size{
-      query->query_->get_est_result_size_variable_nullable(field_name)};
-  *size_off = est_size.fixed_;
-  *size_val = est_size.variable_;
-  *size_validity = est_size.validity_;
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_fragment_num(
-    tiledb_ctx_t* ctx, const tiledb_query_t* query, uint32_t* num) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *num = query->query_->get_written_fragment_num();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_fragment_uri(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t idx,
-    const char** uri) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *uri = query->query_->get_written_fragment_uri(idx).c_str();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_fragment_timestamp_range(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t idx,
-    uint64_t* t1,
-    uint64_t* t2) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  std::tie(*t1, *t2) = query->query_->get_written_fragment_timestamp_range(idx);
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_subarray_t(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    tiledb_subarray_t** subarray) {
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-  ensure_output_pointer_is_valid(subarray);
-  *subarray = make_handle<tiledb_subarray_t>(*(query->query_->subarray()));
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_get_relevant_fragment_num(
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t* relevant_fragment_num) {
-  if (sanity_check(ctx, query) == TILEDB_ERR)
-    return TILEDB_ERR;
-
-  *relevant_fragment_num =
-      query->query_->subarray()->relevant_fragments().relevant_fragments_size();
-
-  return TILEDB_OK;
-}
-
-int32_t tiledb_query_add_update_value(
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* field_name,
-    const void* update_value,
-    uint64_t update_value_size) {
-  // Sanity check
-  if (sanity_check(ctx, query) == TILEDB_ERR) {
-    return TILEDB_ERR;
-  }
-
-  query->query_->add_update_value(field_name, update_value, update_value_size);
-
-  // Success
-  return TILEDB_OK;
-}
-
 /* ****************************** */
 /*          QUERY CONDITION       */
 /* ****************************** */
@@ -2390,285 +1872,6 @@ CAPI_INTERFACE(as_built_dump, tiledb_string_t** out) {
 }
 
 /* ****************************** */
-/*              QUERY             */
-/* ****************************** */
-
-CAPI_INTERFACE(
-    query_alloc,
-    tiledb_ctx_t* ctx,
-    tiledb_array_t* array,
-    tiledb_query_type_t query_type,
-    tiledb_query_t** query) {
-  return api_entry<tiledb::api::tiledb_query_alloc>(
-      ctx, array, query_type, query);
-}
-
-CAPI_INTERFACE(
-    query_get_stats,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    char** stats_json) {
-  return api_entry<tiledb::api::tiledb_query_get_stats>(ctx, query, stats_json);
-}
-
-CAPI_INTERFACE(
-    query_set_config,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_config_t* config) {
-  return api_entry<tiledb::api::tiledb_query_set_config>(ctx, query, config);
-}
-
-CAPI_INTERFACE(
-    query_get_config,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_config_t** config) {
-  return api_entry<tiledb::api::tiledb_query_get_config>(ctx, query, config);
-}
-
-CAPI_INTERFACE(
-    query_set_subarray_t,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const tiledb_subarray_t* subarray) {
-  return api_entry<tiledb::api::tiledb_query_set_subarray_t>(
-      ctx, query, subarray);
-}
-
-CAPI_INTERFACE(
-    query_set_data_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    void* buffer,
-    uint64_t* buffer_size) {
-  return api_entry<tiledb::api::tiledb_query_set_data_buffer>(
-      ctx, query, name, buffer, buffer_size);
-}
-
-CAPI_INTERFACE(
-    query_set_offsets_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint64_t* buffer_offsets,
-    uint64_t* buffer_offsets_size) {
-  return api_entry<tiledb::api::tiledb_query_set_offsets_buffer>(
-      ctx, query, name, buffer_offsets, buffer_offsets_size);
-}
-
-CAPI_INTERFACE(
-    query_set_validity_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint8_t* buffer_validity,
-    uint64_t* buffer_validity_size) {
-  return api_entry<tiledb::api::tiledb_query_set_validity_buffer>(
-      ctx, query, name, buffer_validity, buffer_validity_size);
-}
-
-CAPI_INTERFACE(
-    query_get_data_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    void** buffer,
-    uint64_t** buffer_size) {
-  return api_entry<tiledb::api::tiledb_query_get_data_buffer>(
-      ctx, query, name, buffer, buffer_size);
-}
-
-CAPI_INTERFACE(
-    query_get_offsets_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint64_t** buffer,
-    uint64_t** buffer_size) {
-  return api_entry<tiledb::api::tiledb_query_get_offsets_buffer>(
-      ctx, query, name, buffer, buffer_size);
-}
-
-CAPI_INTERFACE(
-    query_get_validity_buffer,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* name,
-    uint8_t** buffer,
-    uint64_t** buffer_size) {
-  return api_entry<tiledb::api::tiledb_query_get_validity_buffer>(
-      ctx, query, name, buffer, buffer_size);
-}
-
-CAPI_INTERFACE(
-    query_set_layout,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_layout_t layout) {
-  return api_entry<tiledb::api::tiledb_query_set_layout>(ctx, query, layout);
-}
-
-CAPI_INTERFACE(
-    query_set_condition,
-    tiledb_ctx_t* const ctx,
-    tiledb_query_t* const query,
-    const tiledb_query_condition_t* const cond) {
-  return api_entry<tiledb::api::tiledb_query_set_condition>(ctx, query, cond);
-}
-
-CAPI_INTERFACE(query_finalize, tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  return api_entry<tiledb::api::tiledb_query_finalize>(ctx, query);
-}
-
-CAPI_INTERFACE(
-    query_submit_and_finalize, tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  return api_entry<tiledb::api::tiledb_query_submit_and_finalize>(ctx, query);
-}
-
-CAPI_INTERFACE_VOID(query_free, tiledb_query_t** query) {
-  return api_entry_void<tiledb::api::tiledb_query_free>(query);
-}
-
-CAPI_INTERFACE(query_submit, tiledb_ctx_t* ctx, tiledb_query_t* query) {
-  return api_entry<tiledb::api::tiledb_query_submit>(ctx, query);
-}
-
-CAPI_INTERFACE(
-    query_has_results,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    int32_t* has_results) {
-  return api_entry<tiledb::api::tiledb_query_has_results>(
-      ctx, query, has_results);
-}
-
-CAPI_INTERFACE(
-    query_get_status,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_query_status_t* status) {
-  return api_entry<tiledb::api::tiledb_query_get_status>(ctx, query, status);
-}
-
-CAPI_INTERFACE(
-    query_get_type,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_query_type_t* query_type) {
-  return api_entry<tiledb::api::tiledb_query_get_type>(ctx, query, query_type);
-}
-
-CAPI_INTERFACE(
-    query_get_layout,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_layout_t* query_layout) {
-  return api_entry<tiledb::api::tiledb_query_get_layout>(
-      ctx, query, query_layout);
-}
-
-CAPI_INTERFACE(
-    query_get_array,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    tiledb_array_t** array) {
-  return api_entry<tiledb::api::tiledb_query_get_array>(ctx, query, array);
-}
-
-CAPI_INTERFACE(
-    query_get_est_result_size,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size) {
-  return api_entry<tiledb::api::tiledb_query_get_est_result_size>(
-      ctx, query, name, size);
-}
-
-CAPI_INTERFACE(
-    query_get_est_result_size_var,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_off,
-    uint64_t* size_val) {
-  return api_entry<tiledb::api::tiledb_query_get_est_result_size_var>(
-      ctx, query, name, size_off, size_val);
-}
-
-CAPI_INTERFACE(
-    query_get_est_result_size_nullable,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_val,
-    uint64_t* size_validity) {
-  return api_entry<tiledb::api::tiledb_query_get_est_result_size_nullable>(
-      ctx, query, name, size_val, size_validity);
-}
-
-CAPI_INTERFACE(
-    query_get_est_result_size_var_nullable,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    const char* name,
-    uint64_t* size_off,
-    uint64_t* size_val,
-    uint64_t* size_validity) {
-  return api_entry<tiledb::api::tiledb_query_get_est_result_size_var_nullable>(
-      ctx, query, name, size_off, size_val, size_validity);
-}
-
-CAPI_INTERFACE(
-    query_get_fragment_num,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint32_t* num) {
-  return api_entry<tiledb::api::tiledb_query_get_fragment_num>(ctx, query, num);
-}
-
-CAPI_INTERFACE(
-    query_get_fragment_uri,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t idx,
-    const char** uri) {
-  return api_entry<tiledb::api::tiledb_query_get_fragment_uri>(
-      ctx, query, idx, uri);
-}
-
-CAPI_INTERFACE(
-    query_get_fragment_timestamp_range,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t idx,
-    uint64_t* t1,
-    uint64_t* t2) {
-  return api_entry<tiledb::api::tiledb_query_get_fragment_timestamp_range>(
-      ctx, query, idx, t1, t2);
-}
-
-CAPI_INTERFACE(
-    query_get_subarray_t,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    tiledb_subarray_t** subarray) {
-  return api_entry<tiledb::api::tiledb_query_get_subarray_t>(
-      ctx, query, subarray);
-}
-
-CAPI_INTERFACE(
-    query_get_relevant_fragment_num,
-    tiledb_ctx_t* ctx,
-    const tiledb_query_t* query,
-    uint64_t* relevant_fragment_num) {
-  return api_entry<tiledb::api::tiledb_query_get_relevant_fragment_num>(
-      ctx, query, relevant_fragment_num);
-}
-
-/* ****************************** */
 /*          QUERY CONDITION       */
 /* ****************************** */
 
@@ -2737,21 +1940,6 @@ CAPI_INTERFACE(
     int use_enumeration) {
   return api_entry<tiledb::api::tiledb_query_condition_set_use_enumeration>(
       ctx, cond, use_enumeration);
-}
-
-/* ****************************** */
-/*         UPDATE CONDITION       */
-/* ****************************** */
-
-CAPI_INTERFACE(
-    query_add_update_value,
-    tiledb_ctx_t* ctx,
-    tiledb_query_t* query,
-    const char* field_name,
-    const void* update_value,
-    uint64_t update_value_size) {
-  return api_entry<tiledb::api::tiledb_query_add_update_value>(
-      ctx, query, field_name, update_value, update_value_size);
 }
 
 /* ****************************** */
